+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gaindev-ui/eslint-config
 
+## 4.0.0
+
+### Major Changes
+
+- Update code Toast code (styles)
+
 ## 3.0.0
 
 ### Major Changes