+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gaindev-ui/react
 
+## 4.0.0
+
+### minot Changes
+
+- Update code Toast code (styles)
+
 ## 3.0.0
 
 ### Major Changes