+++ packages/ts-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gaindev-ui/ts-config
 
+## 3.0.0
+
+### Major Changes
+
+- Update code Toast code (styles)
+
 ## 2.0.0
 
 ### Major Changes